### 2021-04-06 #3154 @whedon
[PRE REVIEW]: nrCascadeSim - A simulation tool for nuclear recoil cascades resulting from neutron capture
**Submitting author:** @gerudo7 (<a href="http://orcid.org/0000-0001-5406-8367">Kitty Harris</a>)
**Repository:** <a href="https://github.com/villano-lab/nrCascadeSim" target ="_blank">https://github.com/villano-lab/nrCascadeSim</a>
**Version:** 1.0.1
