### 2021-04-11 875cb0d
Add DOIs suggested by whedon bot Checked validity before adding; both are correct.
