+++ paper/references.bib
@@ -9,7 +9,8 @@
     pages = {217--235},
     year = {1998},
     publisher = {International Union of Pure and Applied Chemistry},
-    url = {http://publications.iupac.org/pac/pdf/1998/pdf/7001x0217.pdf}
+    url = {http://publications.iupac.org/pac/pdf/1998/pdf/7001x0217.pdf},
+    doi = {10.1515/iupac.70.0026}
 }
 
 @web{nudat2,
@@ -58,7 +59,8 @@
     issue = {3},
     pages = {1086--1098},
     year = {1991},
-    publisher = {The American Phsyical Society}
+    publisher = {The American Phsyical Society},
+    doi = {10.1103/physrevc.43.1086}
 }
 
 @article{Si,
