### 2021-04-11 dcf2dd6
Update references with DOIs and one with missing authors Could not locate a DOI for Lindhard, though whedon bot didn't mark it missing.
