+++ paper/references.bib
@@ -65,11 +65,12 @@
 
 @article{Si,
     title = {Thermal-neutron capture by silicon isotopes},
-    author = {Raman, S.},
+    author = {Raman, S. and Jurney, E.T. and Lynn, J.E.},
     journal = {Physical Review C},
     year = {1992},
     publisher = {The American Physcial Society},
     pages = {972--983},
     volume = {46},
-    issue = {3}
+    issue = {3},
+    doi = {10.1103/PhysRevC.46.972}
 }
