### 2021-04-14 #3154 @mbobra
> @mbobra I am a co-author of this paper alongside @gerudo7 who submitted the paper. The paper doesn't show up in my JOSS profile but it does in @gerudo7 . Is there a way to make it show up in mine too as a convenience?

I don't think so, I'm sorry! Our system attaches the paper to the submitting author. @openjournals/joss-eics Can you please confirm?
