### 2021-04-15 #3154 @arfon
> I don't think so, I'm sorry! Our system attaches the paper to the submitting author. @openjournals/joss-eics Can you please confirm?

Yeah, I'm afraid the paper can only be associated with a single author in our submission system.
