### 2021-04-15 #3174 @mbobra
👋 @kelseymh, @matthewfeickert Thank you so much for agreeing to review this submission!

Here's what you need:

* Here is the software repository: https://github.com/villano-lab/nrCascadeSim
* The paper describing the software is linked in Whedon's comment above ⬆️
* Here is a detailed description of the review criteria: https://joss.readthedocs.io/en/latest/review_criteria.html#review-items

Please check off the tick boxes (above, in the first comment on this thread) as you go through your review. This is an open review process, and we encourage communication between the reviewers, the submitting author, and the editor (myself). Please feel free to ask me any questions if you have them!
