### 2021-04-16 #3174 @kelseymh
Regarding conflict of interest: @villaa and I are collaborators on the SuperCDMS experiment.  We show up as authors on all SuperCDMS papers over the past several years, and as the simulations software lead in the experiment, I have worked with Anthony directly.  I have no personal concerns about impartially evaluating this nrCascadeSim package (that's part of my resonsibility in the Collaboration).

I will leave the COI checkbox open for now, until I receive a followup from the editors.
