### 2021-04-19 #3174 @matthewfeickert
@mbobra @gerudo7 The current review information has the following for version information

> **Repository:** <a href="https://github.com/villano-lab/nrCascadeSim" target ="_blank">https://github.com/villano-lab/nrCascadeSim</a>
> **Version:** 1.0.1

but the repository has a [tag `1.0.3`](https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.0.3) that was recently released. Unless I am told otherwise, I am going to base the review only on what is available in [tag `v1.0.1`](https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.0.1), but am happy to change the review version if requested.

Also, as there are no version ranges given in the docs on supported versions of ROOT I am going to base my review off of the latest release of ROOT which is [`v6.24.00`](https://root.cern/releases/release-62400/) and conduct all code assessment in the following Docker image: [`atlasamglab/stats-base:root6.24.00`](https://hub.docker.com/r/atlasamglab/stats-base/tags?page=1&ordering=last_updated&name=6.24.00)
