### 2021-04-20 #3174 @mbobra
@matthewfeickert Thanks for looking into this. You raise a really good point.

We allow the submitting authors to make minor changes while the review is in process. It looks like the authors are following semantic versioning, so the changes in `v1.0.3` and `v1.0.2` should not affect your review.  If, however, a reviewer raises a concern during the review process and opens an issue on the submitting author's repo, then the author must address that change and release a new version before we accept the submission for publication. In this case, a new version (1.0.4 or greater) would include the minor changes implemented in `v1.0.3` and `v1.0.2`. Furthermore, the final publication will include your name as a reviewer and therefore indicate your stamp of approval. Let me know if you have any issues with this.

@gerudo7 Please do not make any major changes to `nrCascadeSim` while it is under review.
