### 2021-04-20 9f5551b
Add ROOT version compatibility and default branch for contributors.
