### 2021-04-21 #3174 @gerudo7
>     * [x]  **Contribution and authorship:** Has the submitting author (@gerudo7) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
>       * There is only 1 commit under the username @gerudo7, but it appears that a large number of commits have been contributed by a user of the same first name.
>

Just wanted to cofiirm that this was me and note that I think I was able to fix this. Apparently the email in my .gitconfig file wasn't associated with my GitHub account; adding it to my account seems to have given all the pseudo-account's commits to my real one.
