+++ nrCascadeSim_env.yml
@@ -1,0 +1,23 @@
+name: nrCascadeSim_env 
+dependencies:
+  - python>=3.6
+  - anaconda
+  - numpy
+  - scipy
+  - conda-forge::emcee
+  - pandas
+  - astropy::gammapy
+  - astropy::corner
+  - matplotlib
+  - mpld3
+  - tabulate
+  - h5py
+  - lmfit
+  - ipykernel
+  - nb_conda_kernels
+  - anaconda::pytest
+  - pip
+  - pip:
+    - svgpathtools
+    - bootstrapped 
+    - papermill
