+++ .travis.yml
@@ -1,0 +1,30 @@
+language: python
+python: 3.7
+dist: xenial
+sudo: true
+install:
+  - sudo apt-get update || true
+  - sleep 10
+  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
+  - bash miniconda.sh -b -p $HOME/miniconda
+  - export PATH="$HOME/miniconda/bin:$PATH"
+  - hash -r
+  - conda config --set always_yes yes --set changeps1 no
+  - conda update -q conda || true
+  # Useful for debugging any issues with conda
+  - conda info -a
+
+  - conda env create -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
+  - source activate test-environment
+
+  # Useful for debugging version of emcee 
+  - echo 'Conda searching emcee:'
+    #- conda search emcee
+  - echo 'Python Version:'
+  - echo $TRAVIS_PYTHON_VERSION
+  - conda list |grep emcee 
+
+before_script: cd example-usecase 
+
+script:
+- travis_wait 50 py.test
