### 2021-04-23 234882c
[ANV] added uproot to the Travis environment
