### 2021-04-23 1276187
[ANV] updated the environment yml file
