### 2021-04-23 7d04d8b
[ANV] dont need emcee for this notebook for now
