+++ .travis.yml
@@ -18,11 +18,11 @@ install:
   - source activate test-environment
 
   # Useful for debugging version of emcee 
-  - echo 'Conda searching emcee:'
+  #- echo 'Conda searching emcee:'
     #- conda search emcee
-  - echo 'Python Version:'
-  - echo $TRAVIS_PYTHON_VERSION
-  - conda list |grep emcee 
+  #- echo 'Python Version:'
+  #- echo $TRAVIS_PYTHON_VERSION
+  #- conda list |grep emcee 
 
 before_script: cd example-usecase 
 
