+++ nrCascadeSim_env.yml
@@ -11,3 +11,4 @@ dependencies:
   - pip
   - pip:
     - uproot
+    - papermill
