### 2021-04-30 #3174 @matthewfeickert
Thanks for your submission and software @gerudo7. The following is my first round review of it, which goes into detail (and so isn't just noted in the first post of this Issue) only to try to help make requested revisions more actionable and clear. Unless @mbobra has any additional comments on my review I'll proceed to open up individual GitHub issues that target the topics below in your repo to make it easier to work though them as well. Also, please feel free to ask for clarification on anything as I'd like to make sure that this information is helpful and actionable to you. :+1:

## Review checklist

This review is for [tag `v1.0.1`](https://github.com/villano-lab/nrCascadeSim/releases/tag/v1.0.1) of the project.

   - The repo has an OSI approved license (MIT), but vendors the Msersenne Twister prime algorithm without any LICENSE, which is a problem. All vendored software must be properly licensed. It is also unclear why this is even needed as the [Mersenne Twister 19937 generator is in the C++ Standard Library](https://www.cplusplus.com/reference/random/mt19937/).

   - Out of the number of commits in this project related to the software itself the submitting author @gerudo7 has made the majority of them

   ```
   $ git log --oneline ":(exclude)paper" | wc -l
   50
   $ git log --author=Kitty --oneline ":(exclude)paper" | wc -l
   32
   ```

JOSS asks for a minimum of 3 months of work equivalent effort for the software to be considered. What looks like the bulk of the original code was added in 2019 with development beginning to pick up in late 2020. While

```
git log -- . ":(exclude)paper"
```

is on the short side, it seems from the paper the goal would be to continue development in the future and the statement of need seems to imply that this software could be of particular interest to the dark matter direct detection community, so it seems the work and value is there.

   - No. The installation instructions fail at `make`
```
$ make
g++ -fPIC -c -D__GIT_VERSION=\"v1.0.1\" -IMersenne -Iinc src//isotope_info.c `root-config --cflags --glibs` -L. -Lbin/lib
mv isotope_info.o bin/lib/
mv: cannot move 'isotope_info.o' to 'bin/lib/': Not a directory
make: *** [Makefile:28: bin/lib/isotope_info.o] Error 1
```

Reproducible example:

$ docker run --rm -ti -v $PWD:$PWD -w $PWD atlasamglab/stats-base:root6.24.00
# gcc --version | head -n 1
gcc (Debian 8.3.0-6) 8.3.0
# root --version
ROOT Version: 6.24/00
Built for linuxx8664gcc on Apr 20 2021, 08:31:00
From tags/v6-24-00@v6-24-00
# make
g++ -fPIC -c -D__GIT_VERSION=\"v1.0.1\" -IMersenne -Iinc src//isotope_info.c `root-config --cflags --glibs` -L. -Lbin/lib
mv isotope_info.o bin/lib/
mv: cannot move 'isotope_info.o' to 'bin/lib/': Not a directory
make: *** [Makefile:28: bin/lib/isotope_info.o] Error 1
```

If the `bin/lib` directory is manually created by the user (it should not need to be) then `make` will succeed with a large number of warnings from the vendored distribution of the Mersenne prime algorithm.

   - No, given that it fails to build with given instructions.

   - There are no claims of performance made.

   - To some degree. From the first section of the `README` it is apparently meant to be used

> to simulate energy deposits due to cascading of energy levels following neutron capture.

However, there is no further documentation that is of additional detail as to the relevant physics or the areas of physics that might be interested in using the software.

   - Yes. The software only depends on the `ROOT` C++ library and the `gcc` compiler.

   - No. There is no clear example with context or documented results. The only example is

> The main argument (no prefix) specifies the input file. (example: `levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt`) Making the full example: `./realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt`

which is not in a separate example section or clearly marked. This also fails to indicate that the user should be in the `bin` directory as the executable is not made available to the user in `PATH`.

When all of this is taken into account, the command

./bin/realizeCascades -o output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
```

prints output to `stdout` by default with no additional logging, which makes it quite difficult to digest the relevant information. It would be good if this could be made a non-default option.

I can see that there is a `example-usecase` directory in the repository that is never mentioned in the `README`, but there is no documentation file in the directory itself and the only form of documentation is the notebook, which is a nice visualization but not a demo. There is a Conda `environment.yml`, but one should also include a `requirement.txt`, as there is nothing specific about Conda or any Conda-forge libraries that are needed for the example --- everything can be installed from PyPI.

After creating a CPython 3.8.7 virtual environment I installed the necessary libraries and then tried to run the notebook. It fails on the first cell with

```pytb
<ipython-input-3-0e4771bdf606> in <module>
     19 file = './data/file.root'
     20
---> 21 real_Lind = np.ndarray.flatten(np.asarray(h(file)[0]))
     22 real_Sor = np.ndarray.flatten(np.asarray(h(file,model='Sorenson')[0]))
     23 small_Lind = np.ndarray.flatten(np.asarray(h(file,scalefactor=0.2)[0]))

~/Code/GitHub/JOSS-reviews/nrCascadeSim-review/nrCascadeSim/example-usecase/python/hist.py in histogramable(file, binsize, binmin, binmax, labels, model, method, resolution, val, scalefactor)
     53         x = uproot.open(file)
     54         cas = x['cascade']
---> 55         en = cas.array('E')
     56         en_dep = cas.array('delE')
     57         c_id = cas.array('cid')

AttributeError: 'Model_TTree_v20' object has no attribute 'array'
```

   - No. There is only one command in the library that is available from the CLI API of the program

> The `./realizeCascades` command will run the simulation a specified number of times for a given input file. Currently, this is the only program in this package. It is possible for the user to create additional programs based on teh libraries included, and we may provide additional programs with the library in the future.

Typos in the documentation should be checked for and corrected.

There is some brief description of the custom `.txt` file format (the "levelfiles") but it is not well formatted for Markdown and lacks a full description of the schema and clear examples.

There are also what appear to be "TODO"s written by the authors to themselves

> Check that the below is correct!

that should be removed.

   - No. The repo mentions

> This program is designed to run in a Unix-based system and has been tested on Ubuntu and MacOS.

however there are no tests in the repository (and so no automated tests either). In general, if software is described as being tested on an operating system this should mean that there are automated tests that run the test suite on a specific OS.

   - No. There is no `CONTRIBUTING.md`, no descriptions of using GitHub Issues for bug tracking, no description of using GitHub Discussions, and also no code of conduct.

   - Mostly yes, though there are still assumptions that are too specific to direct detection dark matter physics that should be clarified or removed.

      - **Line 10 - 12**: Goes from introducing tool (should mention is `C++` also) to mentioning what elements are supported but haven't introduced what "support" means so far in the context of the physics.
      - **Line 22**: "the nuclear system have negligible (zero) kinetic energy" reads as saying that because the kinetic energy term is small enough it is dropped from the calculation, but this should be made much more explicit.
      - **Line 47 - 49**: Model parameters of the Lindhard model are introduced without explanation of their values. These should be explained, and it is unclear if they are needed in this paper or if they should be moved to the documentation (probably better).

   Additionally, the paper is too long and includes details that should be included in the documentation. For example, Figure 1 is very nice, but none of it really belongs in the paper and the information in the caption (which should be split out) belongs in the documentation along with the figure. I would keep Figure 2 in the paper, but the section it is in should get refactored as the paper should not serve as documentation for the code. I would recommend taking a look at recently published papers as well as the JOSS example paper on `Gala`: https://joss.theoj.org/papers/10.21105/joss.00388

   - Yes, though the section repeats information in the introduction, so be careful about this in the revision. It could also go into _just_ a little bit more detail.

   - No. This is missing.

   - The writing is overall okay, but could benefit from some revisions.
      - Figure captions should not be as verbose as they currently are. They should describe the figure, but technical information of the figure contents should be described in the main body. However, given the type of paper this is, this information should not be included in the paper and should be in the docs instead.
      - **Line 39**: Is "In+" seems to be a typo.
      - Names of elements should be capitalized.
      - (This did not impact the review, but I thought I would mention this as it was noticeable.) Throughout the paper semicolons are used incorrectly. Though it might sound strange, I would truly recommend Matthew Inman's "The Oatmeal" comic on semicolon usage: https://theoatmeal.com/comics/semicolon
      (I actually routinely use it as a helpful reminder for myself if I'm using semicolons correctly and generally it helps me remember that I shouldn't be using a semicolon at all.)

   - No. ROOT is cited improperly. There are instructions on how to cite ROOT on both [the website](https://root.cern/blog/publications/) and [the GitHub project README](https://github.com/root-project/root/#cite).
