### 2021-04-30 262b2f3
Add mkdir step for bin/lib to prevent make failing.
