### 2021-04-30 33340ef
Add version numbers to dependencies that were previously missing them. Add placeholder for examples.
