+++ README.md
@@ -15,11 +15,11 @@ This program is designed to run in a Unix-based system and has been tested on Ub
 
 Required:
 * CERN's ROOT, which has its own installation instructions here: [https://root.cern/install/](https://root.cern/install/). Intended compatible with all versions; known compatibile with version 6.
-* gcc, which can be installed via the command line (ex: `sudo apt-get install gcc`).
+* gcc, which can be installed via the command line (ex: `sudo apt-get install gcc`). Currently compatible with version 4.4.7 or newer.
 
 Optional:
 * Anaconda or conda, as optionally used in the ROOT installation process: [https://www.anaconda.com/products/individual](https://www.anaconda.com/products/individual)
-* A package for reading *.root files (for Python, one example of this is [uproot](https://pypi.org/project/uproot/) `pip install uproot`)
+* A package for reading *.root files (for Python, one example of this is [uproot](https://pypi.org/project/uproot/) `pip install uproot` - if you choose to use uproot, ensure that you are using version 3.)
 
 ## Steps
 
@@ -34,7 +34,7 @@ Optional:
 # Quick Start
 
 1) If ROOT is not part of the current environment, switch to an appropriate environment (ex: `conda activate name_of_root_env`)
-2) Switch to the appropriate directory (`cd /path/to/directory/nrCascadeSim`)
+2) Switch to the appropriate directory (`cd /path/to/directory/nrCascadeSim/bin`)
 3) ./realizeCascades -n desired-number-of-events -o /path/to/output/file levelfiles/name_of_levelfile.txt
 
 # Instructions for Use
@@ -51,6 +51,8 @@ All three of these arguments are required:
 * `-o` specifies the location of the output file. (example: `-o ~/output.root`)
 * The main argument (no prefix) specifies the input file. (example: `levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt`)
 Making the full example: `./realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt`
+
+(For more detailed examples, see the "Exmaples" section.)
 
 # Levelfile (Input) Format
 
@@ -145,8 +147,12 @@ The ordering of values in the arrays are consistent; that is, the nth entry of `
 The length of each main array should be equal to the number of simulations; that is, 
 if running 10000 events, `n` and `cid` will have lengths of 10000 and the jagged arrays will have first dimensions of length 10000.
 
+# Examples
+
+
+
 # For Contributors
 
 When submitting a pull request, please set the target at `develop`. 
-This allows us to assign version numbers appropriately.
+This allows us to assign version numbers appropriately. 
 Thanks for contributing!
