### 2021-04-30 bf4adbf
Specify that ./realizeCascades must be run from bin dir with ROOT present.
