+++ paper/references.bib
@@ -21,13 +21,20 @@
     url = {https://www.nndc.bnl.gov/nudat2/}
 }
 
-@web{ROOT,
-    title = {ROOT},
-    author = {CERN},
-    publisher = {CERN},
-    year = {2021},
-    url = {https://root.cern/install/}
+@article{ROOT,
+    title = {ROOT - An Object Oriented Data Analysis Framework},
+    author = {Rene Brun and Fons Rademakers},
+    journal = {Nucl. Inst. & Meth. in Phys. Res. A},
+    volume = {389},  
+    pages = {81--86},
+    year = {1997},
+    url = {https://root.cern/}
 }
+%Rene Brun and Fons Rademakers, ROOT - An Object Oriented Data Analysis Framework,
+%Proceedings AIHENP'96 Workshop, Lausanne, Sep. 1996,
+%Nucl. Inst. & Meth. in Phys. Res. A 389 (1997) 81-86.
+%See also "ROOT" [software], Release v6.22/00, 6/15/2020,
+%(Select the right link for your release here: https://zenodo.org/search?page=1&size=20&q=conceptrecid:848818&all_versions&sort=-version).
 
 @article{lindhard,
     title = {Integral Equations Governing Radiation Effects},
