+++ paper/references.bib
@@ -28,19 +28,19 @@
     volume = {389},  
     pages = {81--86},
     year = {1997},
-    note = "See also "ROOT" [software], Release v6.22/00, 6/15/20",
+%    note = "See also "ROOT" [software], Release v6.22/00, 6/15/20",
     note = "See also: \url{https://root.cern/}",
     url = {https://root.cern/}
 }
-@article{ROOT2,
-    title = {ROOT - An Object Oriented Data Analysis Framework},
-    author = {Rene Brun and Fons Rademakers},
-    journal = {Proceedings AIHENP'96 Workshop, Lausanne},
-    year = {1996},
-    note = "See also "ROOT" [software], Release v6.22/00, 6/15/20",
-    note = "See also: \url{https://root.cern/}",
-    url = {https://root.cern/}
-}
+%@article{ROOT2,
+%    title = {ROOT - An Object Oriented Data Analysis Framework},
+%    author = {Rene Brun and Fons Rademakers},
+%    journal = {Proceedings AIHENP'96 Workshop, Lausanne},
+%    year = {1996},
+%    note = "See also "ROOT" [software], Release v6.22/00, 6/15/20",
+%    note = "See also: \url{https://root.cern/}",
+%    url = {https://root.cern/}
+%}
 %Rene Brun and Fons Rademakers, ROOT - An Object Oriented Data Analysis Framework,
 %Proceedings AIHENP'96 Workshop, Lausanne, Sep. 1996,
 %Nucl. Inst. & Meth. in Phys. Res. A 389 (1997) 81-86.
