+++ paper/references.bib
@@ -28,10 +28,10 @@
     volume = {389},  
     pages = {81--86},
     year = {1997},
-%    note = "See also "ROOT" [software], Release v6.22/00, 6/15/20",
     note = "See also: \url{https://root.cern/}",
     url = {https://root.cern/}
 }
+%    note = "See also "ROOT" [software], Release v6.22/00, 6/15/20",
 %@article{ROOT2,
 %    title = {ROOT - An Object Oriented Data Analysis Framework},
 %    author = {Rene Brun and Fons Rademakers},
