### 2021-04-30 1918bce
Fix clarity issues, missing \`, and leftover TODOs that had been finished.
