### 2021-05-01 9e92fbf
Attempt to add basic test that program makes and runs.
