+++ .travis.yml
@@ -1,7 +1,6 @@
 language: python
 python: 3.7
 dist: xenial
-sudo: true
 install:
   - sudo apt-get update || true
   - sleep 10
@@ -24,7 +23,16 @@ install:
   #- echo $TRAVIS_PYTHON_VERSION
   #- conda list |grep emcee 
 
+  - conda create -c conda-forge --name root_env root
+  - conda activate root_env
+  - make
+
 before_script: cd example-usecase 
 
 script:
 - travis_wait 50 py.test
+- cd ../bin
+- ./realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
+
+
+
