### 2021-05-01 c6b1396
Fix ROOT to install to existing conda env instead of new.
