### 2021-05-01 b458e37
Attempt to fix environment issues with ROOT by using `conda activate` instead of `source activate`
