### 2021-05-01 bdb7d35
Could not conda activate; trying conda init first If this doesn't work will revert back to source activate and change root installation.
