+++ .travis.yml
@@ -14,6 +14,7 @@ install:
   - conda info -a
 
   - conda env create -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
+  - conda init bash
   - conda activate test-environment
 
   # Useful for debugging version of emcee 
