### 2021-05-01 3bd9df7
Revert to source activate See if it's somehow exiting the conda environment and just needs to jump back in during test?
