### 2021-05-01 185a9a3
Add root to environment. This should fix auto-build/test.
