+++ nrCascadeSim_env.yml
@@ -1,4 +1,7 @@
 name: nrCascadeSim
+channels:
+  - conda-forge
+  - default
 dependencies:
   - python>=3.6
   - anaconda
@@ -12,3 +15,4 @@ dependencies:
   - pip:
     - uproot==3.12.0
     - papermill
+  - root
