### 2021-05-01 a2e8974
Remove extra conda activate command Hopefully issue is interference with source activate command.
