+++ .travis.yml
@@ -31,7 +31,6 @@ before_script: cd example-usecase
 script:
 - travis_wait 50 py.test
 - cd ../bin
-- conda activate test-environment
 - ./realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
 
 
