### 2021-05-01 cbe5125
Create CONTRIBUTING.md file And remove "for contributors" section from README
