+++ CONTRIBUTING.md
@@ -1,0 +1,20 @@
+# Code of Conduct
+
+
+# Standards
+
+
+
+
+# Pull Requests
+
+When submitting pull requests, 
+please default to submitting to the `develop` branch and never submit directly to the `master` branch. 
+This allows us to correctly manage versioning. 
+If you submit a pull request directly to the `master` branch, you will be asked to change the target to `develop`.
+
+# GitHub Issues
+
+
+
+# GitHub Discussions
