### 2021-05-04 #3174 @gerudo7
@matthewfeickert @mbobra Thank you for your comments, we have begun addressing them in a new release branch (1.0.4). Would you prefer to open issues on GitHub or should we go ahead and do so?
