+++ .travis.yml
@@ -30,8 +30,8 @@ before_script: cd example-usecase
 
 script:
 - travis_wait 50 py.test
-- cd ../bin
-- ./realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
+#- cd ../bin
+#- ./realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
 
 
 
