### 2021-05-04 #3174 @matthewfeickert
> @matthewfeickert @mbobra Thank you for your comments, we have begun addressing them in a new release branch (1.0.4). Would you prefer to open issues on GitHub or should we go ahead and do so?

I can open GitHub Issues later. I would highly recommend that instead of just trying to fix everything by only committing directly to one branch that you adopt a PR based GitFlow development strategy so that there is a `1:1` relationship between a GitHub Issue and a PR that resolves that Issue. It will be much easier as you're adding tests to know if you've resolved the issue properly or not if your CI passes on the PR before you merge it.
