### 2021-05-04 #3174 @villaa
> I would highly recommend that instead of just trying to fix everything by only committing directly to one branch that you adopt a PR based GitFlow development strategy so that there is a 1:1 relationship between a GitHub Issue and a PR that resolves that Issue.

I am new to PR based GitFlow development, I would appreciate any references you have.
