### 2021-05-04 #3174 @villaa
> Regarding conflict of interest: @villaa and I are collaborators on the SuperCDMS experiment. We show up as authors on all SuperCDMS papers over the past several years, and as the simulations software lead in the experiment, I have worked with Anthony directly. I have no personal concerns about impartially evaluating this nrCascadeSim package (that's part of my resonsibility in the Collaboration).
>
> I will leave the COI checkbox open for now, until I receive a followup from the editors.

@mbobra can you respond regarding @kelseymh 's COI? If there is a COI I would be happy to help to try to identify a new reviewer. Thank you for your time.
