### 2021-05-04 villano-lab/nrCascadeSim#2 @gerudo7
Installation fails at make
From **matthewfeickert** in `openjournals/joss-reviews/issues/3174` on 04/29/21:
> If the `bin/lib` directory is manually created by the user (it should not need to be) then make will succeed with a large number of warnings from the vendored distribution of the Mersenne prime algorithm.

**Description:** If following the installation directions given, the command `make` fails because the `bin/lib` directory is not present. There are also a large number of warnings (these will be submitted as a separate issue).

**Solution:** Have `make` create the `bin/lib` directory.

**Testing:** No automated testing was performed for the `make` process.
