### 2021-05-04 villano-lab/nrCascadeSim#2 @gerudo7
Issue resolved in 8beac0b6418b029f53bb7e25b68f7a8c2a4f2d69 on branch `release/1.0.4`. *This commit is part of 1.0.4 and was submitted before switching to PR-based issue resolution; after v1.0.4 is released, issues will be resolved with pull requests.*

**Solution:** Have `make` create the `bin/lib` directory.

**Resolution:** The `make` command now succeeds without the user needing to create the `bin/lib` directory.

**Testing:** Automated testing for `make` process added in 9e92fbf228cfe77f396fc639c8f296e9d1cb6d33 and build passing as of 60440c3b4f1cb29276b52b505f01c9f0e44fa629.

Issue will be marked resolved when v1.0.4 is released.
