### 2021-05-04 villano-lab/nrCascadeSim#3 @gerudo7
Clarity Issue - Statement of Need in README
From **matthewfeickert** in `openjournals/joss-reviews#3174` on 04/29/21 [~~Formatting changed~~ EDIT: fixed quote formatting when posting but forgot to remove note.]:
> >**A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>
>To some degree. From the first section of the README it is apparently meant to be used
>> to simulate energy deposits due to cascading of energy levels following neutron capture.
>
> However, there is no further documentation that is of additional detail as to the relevant physics or the areas of physics that might be interested in using the software.

**Description:** No documentation specifies what physics or areas of physics the code is useful for.

**Solution:** Elaborate on this part of the README to address what the relevant physics/area of physics is.

**Testing:** N/A (Documentation issue)
