### 2021-05-04 villano-lab/nrCascadeSim#3 @gerudo7
Worked on this in c39829a9412ecc21ff813cfe2be9549736884a82 and d67073d4405c0674ae45b8b47e751322547ff52f on branch `release/1.0.4`. Will revisit after releasing v1.0.4.
