### 2021-05-04 villano-lab/nrCascadeSim#4 @gerudo7
ROOT Citation incorrect
From **matthewfeickert** in `openjournals/joss-reviews#3174` on 04/29/21:
> ROOT is cited improperly. There are instructions on how to cite ROOT on both the website and the GitHub project README.

**Description:** ROOT citation is incorrect.

**Solution:** Follow ROOT citation outlined on their website/GitHub README file.

**Testing:** N/A (Paper issues)
