### 2021-05-04 villano-lab/nrCascadeSim#4 @gerudo7
Assigned @villaa along with myself because he has started on this in 4d28a9001bec340712e9750efff443f761bc1678.
From **villaa** on 04/30/21:
> couldn't figure out how to get both the workshop and NIMA paper from here: https://github.com/root-project/root/#cite so did just the NIMA paper. Also, the "notes" field from our bibtex isn't showing up so I don't know where to do the note about the version: 6.22/00 6/15/20.
