### 2021-05-04 villano-lab/nrCascadeSim#5 @gerudo7
Mersenne Twister - License missing!
From **matthewfeickert** in `openjournals/joss-reviews#3174` on 04/29/21:
> The repo has an OSI approved license (MIT), but vendors the Msersenne Twister prime algorithm without any LICENSE, which is a problem. All vendored software must be properly licensed. It is also unclear why this is even needed as the Mersenne Twister 19937 generator is in the C++ Standard Library.

**Description:** No license is provided for Mersenne Twister, which is vendored.

**Solution:** Either fix the license issue by providing a separate license for MT, or use the standard C++ MT library.

**Testing:** N/A (License issue)
