### 2021-05-04 villano-lab/nrCascadeSim#5 @gerudo7
Due to another issue:
From **matthewfeickert** in `openjournals/joss-reviews#3174` on 04/29/21:
> `make` will succeed with a large number of warnings from the vendored distribution of the Mersenne prime algorithm.

The ideal solution is likely to use the standard MT algorithm provided with C++ 11 or newer. However, previously this code has been used with very old versions of C++ (10 or older), so switching methods would be a breaking change. For the time being, the oldest version of gcc known to work has been specified in the README (commit 33340efbfc311a2eeb17466330e2124d8955c1e9 on release/1.0.4) so it will hopefully be clear, once the change is made, why it was a breaking change.

Assigning @villaa because he is more familiar with the vendored MT implementation as well as myself as I am more familiar with the standard C++ implementation.
