### 2021-05-04 villano-lab/nrCascadeSim#6 @gerudo7
example-usecase/Yields_and_Resolutions.ipynb - First cell fails to run.
From **matthewfeickert** in `openjournals/joss-reviews#3174` on 04/29/21:
> After creating a CPython 3.8.7 virtual environment I installed the necessary libraries and then tried to run the notebook. It fails on the first cell with
```python
<ipython-input-3-0e4771bdf606> in <module>
     19 file = './data/file.root'
     20
---> 21 real_Lind = np.ndarray.flatten(np.asarray(h(file)[0]))
     22 real_Sor = np.ndarray.flatten(np.asarray(h(file,model='Sorenson')[0]))
     23 small_Lind = np.ndarray.flatten(np.asarray(h(file,scalefactor=0.2)[0]))

~/Code/GitHub/JOSS-reviews/nrCascadeSim-review/nrCascadeSim/example-usecase/python/hist.py in histogramable(file, binsize, binmin, binmax, labels, model, method, resolution, val, scalefactor)
     53         x = uproot.open(file)
     54         cas = x['cascade']
---> 55         en = cas.array('E')
     56         en_dep = cas.array('delE')
     57         c_id = cas.array('cid')

AttributeError: 'Model_TTree_v20' object has no attribute 'array'
```

**Description:** First cell of provided notebook `example-usecase/Yields_and_Resolutions.ipynb` fails to run, throwing an error, when using the provided

**Solution:** Make the cell run with the provided environment specifications.

**Testing:** This notebook is tested automatically on branch `release/1.0.4` (see `.travis.yml` which references `example-usecase/test_local.py`).
