### 2021-05-04 villano-lab/nrCascadeSim#6 @gerudo7
Uproot version specified by @villaa in `example-usecase/environment.yml` in commit 57426820e601a917c84bfae92441e46fc636038d as the issue seems to be a result of a breaking change in uproot going from version 3 to 4. I will revisit whether this is the optimal solution after release 1.0.4, as any other fix will likely require breaking changes.
