### 2021-05-04 villano-lab/nrCascadeSim#7 @gerudo7
Examples need more clarity and elaboration
From **matthewfeickert** in `openjournals/joss-reviews#3174` on 04/29/21:
>> Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
>   No. There is no clear example with context or documented results. The only example is
>>    The main argument (no prefix) specifies the input file. (example: `levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt`) Making the full example: `./realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt`
>
> which is not in a separate example section or clearly marked. This also fails to indicate that the user should be in the bin directory as the executable is not made available to the user in PATH.

> I can see that there is a example-usecase directory in the repository that is never mentioned in the README, but there is no documentation file in the directory itself and the only form of documentation is the notebook, which is a nice visualization but not a demo.

**Description:** There are no solid examples, and the examples that are present are hard to find.

**Solution:**
- [ ] Create examples with context and documented results, or expand on existing example to give context and document the results.
- [ ] Create an examples section or otherwise mark where the examples can be found so it is easy for the reader to locate them.
- [ ] Mention the example-usecase directory in the README so that users know where to find it.
- [ ] Specify the location from which the program must be run or add the program to PATH so that it is available everywhere.

**Testing:** N/A (Documentation)
