### 2021-05-04 villano-lab/nrCascadeSim#7 @gerudo7
Resolved the more minor parts; need to work on the actual examples themselves.

- [ ] **Create examples with context and documented results, or expand on existing example to give context and document the results.**
I will address this after 1.0.4 so we can switch to using pull requests.
- [x] **Create an examples section or otherwise mark where the examples can be found so it is easy for the reader to locate them.**
Created an example section, though it's a bit of a skeleton for now, in b5fb7e7701b09951ffa160e035fbfd3046331350.
- [x] **Mention the example-usecase directory in the README so that users know where to find it.**
Fixed in b458e37de2564f3fe27174653e33974217b62370.
- [x] **Specify the location from which the program must be run or add the program to PATH so that it is available everywhere.**
Fixed in bf4adbf4ae60904383fc5d2613d457099ffbf289q, then executable default location changed, then fixed again in 5e03bd4ba6148641ce1a3500a8db8a9fa415fcab.
- [ ] **[Added] Make a demo, not just a visualization, for realizeCascades.**
