### 2021-05-04 villano-lab/nrCascadeSim#8 @gerudo7
Functionality Documentation - Clarity Issues
From **matthewfeickert** in `openjournals/joss-reviews#3174` on 04/29/21
> Typos in the documentation should be checked for and corrected.
>There is some brief description of the custom .txt file format (the "levelfiles") but it is not well formatted for Markdown and lacks a full description of the schema and clear examples.
>There are also what appear to be "TODO"s written by the authors to themselves
>>  Check that the below is correct!
>
>that should be removed.

**Description:** There are issues with spelling/grammar or other typose, there are leftover "TODO"s, and the the levelfile documentation is unclear.

**Solution:** Check spelling/grammar and leftover "TODO"s and work on the writing and formatting for levelfile documentation.

**Testing:** N/A (Documentation)
