### 2021-05-04 villano-lab/nrCascadeSim#8 @gerudo7
I have checked the README over for typos already and found a few, but I'm not sure I found all of them. @villaa would you be willing to help me look for them? Thanks!

- [ ] **Check for spelling/grammar/other typos**
I have done this but I am not confident that I found all of them - will treat this as an ongoing task until the other parts are completed.
- [X] **Clear leftover "TODO"s.**
I checked for these and found two - both have been completed and removed.
- [ ] **Work on clarity of both writing and formatting of the levelfiles. Provide schema and clear examples.**
I will start on this rework once we finish 1.0.4 so I can use the pull request workflow instead.
