### 2021-05-04 villano-lab/nrCascadeSim#9 @gerudo7
CONTRIBUTING.md Missing
From **matthewfeickert** in `openjournals/joss-reviews#3174` on 04/29/21
> 	There is no CONTRIBUTING.md, no descriptions of using GitHub Issues for bug tracking, no description of using GitHub Discussions, and also no code of conduct.

**Description:** CONTRIBUTING.md or something equivalent is non-existent.

**Solution:** Create a clear and complete document outlining use of GitHub issues, description of using GitHub discussions, and code of conduct.

**Testing:** N/A
