### 2021-05-04 villano-lab/nrCascadeSim#9 @gerudo7
There is now a placeholder CONTRIBUTING.md and **villaa** and I are collaborating on an outline for its contents.
