### 2021-05-04 #3174 @gerudo7
(Sorry about that ^ - I forgot that it would link. I prevented the linking on all the other issues.)

I went ahead and opened issues only for things we'd already started addressing so that I could try to keep our progress organized. I'm going to get 1.0.4 released here soon so we can switch over to the PR-based strategy.
