### 2021-05-05 #3174 @matthewfeickert
> @matthewfeickert I am new to PR based GitFlow development, I would appreciate any references you have.

Sure. If you search "gitflow" you'll come up with a plethora of resources, some which will give you conflicting advice on how to go about development with regards to the need for `main` and `develop` branches vs. just using a `main` branch (which is more common these days in the era of abundant CI &mdash; note I am not a historian of software development best practices, so downweight that as you will). Example articles that demonstrate this conflict:

- GitHub's ["Understanding the GitHub flow"](https://guides.github.com/introduction/flow/)
- Atlassian's ["Gitflow Workflow"](https://www.atlassian.com/git/tutorials/comparing-workflows/gitflow-workflow)

If you're not familiar with using PRs for development though I'd recommend getting comfortable with that by starting off with [GitHub's intro guide](https://guides.github.com/activities/hello-world/) and also watching [Anthony Sottile's step-by-step video on how to make a PR](https://youtu.be/cysuuUtbC6E).

Note also that GitHub allows [PRs to close Issues automatically on merge with the "resolves", "closes", and "fixes" keywords](https://docs.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword) (among others), so it is very easy to setup a `1:1` relationship.

I would encourage you to look around at many of the well known projects that you use and love and check our their development styles as well as looking at projects that have been published in JOSS.

As a reviewer though I am probably already overstepping my role here by telling you how I would go about doing your development (sorry, not trying to overstep my bounds here or be annoying) so I'll probably stop here. I'll open up remaining Issues if there are any left that @gerudo7 hasn't hit already but I don't think I should be active in giving further feedback on them while you develop and revise &mdash; I'll defer to @mbobra's guidance and advice here RE: everything I just said.
