### 2021-05-05 villano-lab/nrCascadeSim#10 @matthewfeickert
[JOSS Review]: Add testing with all described dependencies with continuous integration
In https://github.com/openjournals/joss-reviews/issues/3174:

> **Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?
>
> No. The repo mentions
> > This program is designed to run in a Unix-based system and has been tested on Ubuntu and MacOS.
>
> however there are no tests in the repository (and so no automated tests either). In general, if software is described as being tested on an operating system this should mean that there are automated tests that run the test suite on a specific OS.

Any operating system or version range of dependencies that are stated to be supported should have the unit test suite run on them in CI. There are a large number of CI services to choose from today and a large number of them support macOS and all of them support Linux. They further support build matrices that allow for testing different runtime versions of dependencies at once. If a wide range of dependencies are claimed to be supported it isn't necessarily required to test ever single one (though a good idea) but the oldest supported and a reasonable selection of modern versions of those dependencies should be tested.

The above is in my mind what constitutes a good test suite, but I'll link to the [JOSS review criteria is for tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests), as my own personal opinions are not the review requirements.
