### 2021-05-05 villano-lab/nrCascadeSim#9 @matthewfeickert
While not required in the [JOSS review criteria for Community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) a Code of Conduct is a great thing to have.
