### 2021-05-05 villano-lab/nrCascadeSim#11 @matthewfeickert
[JOSS Review]: Revise JOSS paper to be shorter and not documentation
From https://github.com/openjournals/joss-reviews/issues/3174:

> - [ ] **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
>
>    - Mostly yes, though there are still assumptions that are too specific to direct detection dark matter physics that should be clarified or removed.
>
>       - **Line 10 - 12**: Goes from introducing tool (should mention is `C++` also) to mentioning what elements are supported but haven't introduced what "support" means so far in the context of the physics.
>       - **Line 22**: "the nuclear system have negligible (zero) kinetic energy" reads as saying that because the kinetic energy term is small enough it is dropped from the calculation, but this should be made much more explicit.
>       - **Line 47 - 49**: Model parameters of the Lindhard model are introduced without explanation of their values. These should be explained, and it is unclear if they are needed in this paper or if they should be moved to the documentation (probably better).
>
>    Additionally, the paper is too long and includes details that should be included in the documentation. For example, Figure 1 is very nice, but none of it really belongs in the paper and the information in the caption (which should be split out) belongs in the documentation along with the figure. I would keep Figure 2 in the paper, but the section it is in should get refactored as the paper should not serve as documentation for the code. I would recommend taking a look at recently published papers as well as the JOSS example paper on `Gala`: https://joss.theoj.org/papers/10.21105/joss.00388
>
> - [x] **A statement of need:** Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
>
>    - Yes, though the section repeats information in the introduction, so be careful about this in the revision. It could also go into _just_ a little bit more detail.
>
  > - [ ] **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
  >
  >    - No. This is missing.
  >
> - [ ] **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
>
>    - The writing is overall okay, but could benefit from some revisions.
>       - Figure captions should not be as verbose as they currently are. They should describe the figure, but technical information of the figure contents should be described in the main body. However, given the type of paper this is, this information should not be included in the paper and should be in the docs instead.
>       - **Line 39**: Is "In+" seems to be a typo.
>       - Names of elements should be capitalized.
>       - (This did not impact the review, but I thought I would mention this as it was noticeable.) Throughout the paper semicolons are used incorrectly. Though it might sound strange, I would truly recommend Matthew Inman's "The Oatmeal" comic on semicolon usage: https://theoatmeal.com/comics/semicolon
>       (I actually routinely use it as a helpful reminder for myself if I'm using semicolons correctly and generally it helps me remember that I shouldn't be using a semicolon at all.)

For some additional guidance check out also the [JOSS review critera for the JOSS paper](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper).
