### 2021-05-05 #3174 @mbobra
@kelseymh If you're close collaborators, I think it violates the COI. If you appear together on papers or are equal parts of a very large collaboration, then that's okay. Does this help, or is it still a bit vague?

cc @villaa
