### 2021-05-05 #3174 @kelseymh
@mbobra CDMS is a "large" collaboration by other science's standards, but we're fairly small in high energy physics, only about 100 people.  Anthony and I have worked together as part of the overal Simulations Working Group within CDMS, but we have only directly collaborated (in the vernacular) on getting the simulation geometry correct for the detector.  In particular, I wasn't involved at all on this project, and wasn't even aware that it existed in any form other than "oh, we do some reweighting of events in our simulation to get the decay channels correct."

My personal feeling is that I don't have a conflict in honestly reviewing and critiquing Anthony's work (been there, done that before :-) ).  However, it's the _appearance_ that's important, and I'm comfortable with whatever you all decide.
