### 2021-05-06 #3174 @mbobra
@kelseymh You can proceed as a reviewer. I very much appreciate your open disclosure. This is a common situation for either large scientific collaborations or large open source scientific software packages (that is, packages with many contributors) and there is precedent for this level of collaboration between a reviewer and submitting author.
