### 2021-05-12 #3174 @villaa
@mbobra can the editor team please comment on the Mersenne Twister license issue brought up by @matthewfeickert. I am happy to address it and we have an issue for it now but I want to clarify what is needed for publication to proceed. Ultimately I would like to move to the C++ standard library implementation but that may actually force a compiler compatibility upgrade due to C++11 standard being needed.

If the code is now included in the C++11 standard library it seems to me the license for the code must be quite permissive, and it probably was even on older versions I would imagine.

**Main question:** what can we include in the Mersenne folder to show the license is appropriate for our use and avoid migration to the C++ standard library form of the code for now (before publication of this article) but keep it on our issue list for upgrade after publication in the future?
