### 2021-05-12 #3174 @matthewfeickert
> If the code is now included in the C++11 standard library it seems to me the license for the code must be quite permissive, and it probably was even on older versions I would imagine.

The permissibility of a license isn't the issue, but rather that code that wasn't written by the authors is being vendored with no license. Unless code explicitly uses [the Unlicense](https://en.wikipedia.org/wiki/Unlicense) or states that it is in the public domain it needs a license for anyone to be able to use it.

This is easy to fix. If you look in the vendored code for something identifiable, like [this line](https://github.com/villano-lab/nrCascadeSim/blob/a7a96e1e0554a9ed37cd2aeecea36ad9534026b7/Mersenne/MersenneTwister.html#L54):

```html
<p>On my system, a Pentium III running Linux at 500 MHz, the performance test gives the following results for generation of random integers:
```

a Google search shows that this was code copied from [Rick Wagner's website](https://faculty.washington.edu/wijsman/progdists/gigi/software/GIGI/GIGI_v1.04/Mersenne/MersenneTwister.html) as part of [his original GIGI code](https://faculty.washington.edu/wijsman/progdists/gigi/software/GIGI/GIGI.html). I can see from there (and poking around in the directory structure of https://faculty.washington.edu/wijsman/progdists/gigi/software/GIGI/) that it is licensed under LGPL.

> Ultimately I would like to move to the C++ standard library implementation but that may actually force a compiler compatibility upgrade due to C++11 standard being needed.

If the code isn't compatible with C++11 that should be noted in the requirements.
