### 2021-05-12 #3174 @villaa
@matthewfeickert

> > Also, the software license was changed from the GNU Lesser General Public License to a BSD license, making commercial use of this software more convenient.
>
> This checks out as we see it also [pop up under BSD on `nixbit`](https://nixbit.com/software/mersennetwister-review/) (from the same search).

Thank you. What I need to know from here is *how* to change our code. Would it be by going to one of the links you list, re-downloading the code and including a license file?
