### 2021-05-12 #3174 @villaa
> > Ultimately I would like to move to the C++ standard library implementation but that may actually force a compiler compatibility upgrade due to C++11 standard being needed.
>
> If the code isn't compatible with C++11 that should be noted in the requirements.

This was noted in the new readme file.
