### 2021-05-12 #3174 @matthewfeickert
> Would it be by going to one of the links you list, re-downloading the code and including a license file?

I'd just throw a BSD license file with Rick's name on in it and the year the code was copied and call it good. You can get a copy of the BSD license template from a few places online, but easiest is Wikipedia: https://en.wikipedia.org/wiki/BSD_licenses#4-clause_license_(original_%22BSD_License%22)
