### 2021-05-13 villano-lab/nrCascadeSim#12 @gerudo7
realizeCascades is overly verbose
From **matthewfeickert** in `openjournals/joss-reviews#3174` on 04/29/21
> the command
`./bin/realizeCascades -o output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt`
prints output to stdout by default with no additional logging, which makes it quite difficult to digest the relevant information. It would be good if this could be made a non-default option.

**Description:** It is difficult to determine what information is relevant because too much is printed.
**Solution:** Reduce the amount of information printed to stdout when running realizeCascades
**Testing:** N/A
