### 2021-05-13 villano-lab/nrCascadeSim#6 @gerudo7
Marking this as resolved as the notebook now runs correctly in the specified environment. May update the notebook in a later version.
