### 2021-05-13 villano-lab/nrCascadeSim#5 @gerudo7
For now we are going to resolve this by adding the license; we can make the transition to the C++ 11 implementation later.
