+++ example-usecase/requirements.txt
@@ -1,0 +1,7 @@
+numpy
+scipy
+matplotlib
+ipykernel
+pytest
+uproot==3.12.0
+papermill
