### 2021-05-14 e80816c
Add debug flag to conda env create to track down travis-CI issue
