### 2021-05-14 77bed96
Add strict channel priority for conda Supposedly can help with hanging at environment solve.
