+++ .travis.yml
@@ -13,6 +13,7 @@ install:
   # Useful for debugging any issues with conda
   - conda info -a
 
+  - conda config --set channel_priority strict
   - conda env create --debug -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
   - source activate test-environment
 
