+++ .travis.yml
@@ -1,5 +1,5 @@
 language: python
-python: 3.7
+python: 3.7.9
 dist: xenial
 install:
   - sudo apt-get update || true
