+++ nrCascadeSim_env.yml
@@ -15,4 +15,4 @@ dependencies:
   - pip:
     - uproot==3.12.0
     - papermill
-  - root
+#  - root
