### 2021-05-16 57785a5
Remove conda-forge from channels. Seems to be the source of the speed issues. ROOT will need to be installed another way - try snap.
