### 2021-05-16 e51c2da
Remove strict channel priority Not necessary on sikngle-channel and interferes with finding packages.
