### 2021-05-16 81ab4b3
Switch back from snap install of root to conda Adding root to env manually rather than from env file
