+++ .travis.yml
@@ -17,9 +17,10 @@ install:
   - conda env create --debug -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
   - source activate test-environment
 
-  - sudo apt-get install snap
-  - sudo snap install root-framework
-  - sudo run root-framework
+  # Snap installation of root
+  #- sudo apt-get install snap
+  #- sudo snap install root-framework
+  #- sudo run root-framework
 
   # Useful for debugging version of emcee 
   #- echo 'Conda searching emcee:'
