### 2021-05-16 2432ac1
Attempt to re-add the make and run commands. If this works then hotfix is done!
