+++ .travis.yml
@@ -30,12 +30,12 @@ install:
   #- conda list |grep emcee 
 
   - conda install root -c conda-forge
-#  - make
+  - make
 
 before_script: cd example-usecase 
 
 script:
 - travis_wait 50 py.test
-#- cd ../bin
-#- bin/realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
+- cd ../bin
+- bin/realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
 # run some kind of script that checks the output.
