### 2021-05-16 86c6c19
Correct commands to match directory structure.
