+++ .travis.yml
@@ -36,6 +36,6 @@ before_script: cd example-usecase
 
 script:
 - travis_wait 50 py.test
-- cd ../bin
+- cd ../
 - bin/realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
 # run some kind of script that checks the output.
