### 2021-05-17 3f29f60
Attempt to fully patch out conda-forge Root is a conda-forge package so not sure why this works, but it works locally? Build for some reason failed once and passed once on the same commit - but think this is the issue.
