+++ .travis.yml
@@ -34,7 +34,7 @@ install:
   # Get ROOT via binary
   - wget https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
   - tar -xzvf root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
-  - sudo apt-get install make cmake gcc-c gcc binutils libX11-devel libXpm-devel libXft-devel libXext-devel openssl-devel
+  - sudo apt-get install make cmake g++ gcc binutils libx11-dev libxpm-dev libxft-dev libxext-dev openssl libssl-dev
   - source root/bin/thisroot.sh
 
   - make
