### 2021-05-17 villano-lab/nrCascadeSim#14 @gerudo7
`example-usecase` directory could use requirement.txt in addition to environment.yml
From **matthewfeickert** in `openjournals/joss-reviews#3174` on 04/29/21:
>There is a Conda environment.yml, but one should also include a requirement.txt, as there is nothing specific about Conda or any Conda-forge libraries that are needed for the example --- everything can be installed from PyPI.

**Description:** There is an environment.yml, but since it is possible, there should also be a requirement.txt.
**Solution:** Add a working requirement.txt.
**Testing:** There should likely be a secondary CI test of the notebook using the pip-based installation to ensure that the file is configured correctly.
