+++ .travis.yml
@@ -15,6 +15,9 @@ install:
 
   #- conda config --set channel_priority strict
   - conda env create --debug -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
+  - conda env create -n piptest
+  - source activate piptest
+  - pip install -r example-usecase/requirements.txt
   - source activate test-environment
 
   # Snap installation of root
@@ -45,4 +48,6 @@ script:
 - travis_wait 50 py.test
 - cd ../
 - bin/realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
+- source activate piptest
+- travis_wait 50 py.test
 # run some kind of script that checks the output.
