+++ .travis.yml
@@ -48,6 +48,8 @@ script:
 - travis_wait 50 py.test
 - cd ../
 - bin/realizeCascades -n 100000 -o ~/output.root levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
+
+after_script:
 - source activate piptest
-- travis_wait 50 py.test
+- python example-usecase/Yields_and_Resolutions.py
 # run some kind of script that checks the output.
