### 2021-05-18 da139c6
Run python script from correct directory. Update README to specify this must be done. Also put this test in the `script` section instead of `after_script` to get it to flag correctly.
