### 2021-05-18 351fd06
Update README.md to mention the new python script.
