### 2021-05-18 3b73998
Clean up testing by adding nbconvert to environment.
