+++ example-usecase/Yields_and_Resolutions.py
@@ -1,124 +1,0 @@
-#A non-notebook version of Yields_and_Resolutions.ipynb
-#Mostly for testing reasons.
-
-#Description
-'''
-Applying Yields and Resolutions
-
-This notebook serves two purposes. The first is to act as a test; you should be able to roughly reproduce the below graphic by running this notebook. The second is to serve as an example usecase. Below, we want to emulate a detector that we know follows either the Lindhard or Sorenson model for ionization energy and has a resolution that affects where the data lands; we are able to apply both the yields and resolutions to the data in order to get a more realistic model for the detector's output.
-
-Generating a *.root file for this notebook¶
-
-The file used to generate the above plot is already present. However, if you would like to generate your own file for comparison, you can replace it as follows. In the top-level nrCascadeSim directory, after compiling (and activating any necessary environments if applicable), run:
-
-./realizeCascades -n 10000 -o test-example/data/file.root levelfiles/Si28_ngam_all_cascades.txt
-
-(Note that due to the randomness of the output, some variation is expected if you replace the file.)
-Notes
-
-You may encounter some runtime warnings - these are expected.
-If you have a *.root file you want to call saved to a different location than mentioned in the instructions above, be sure to change line 21 to point to the correct location.
-Please allow sufficient time to run this notebook; for very large root files it could take up to an hour. The provided file should only take a few minutes.
-'''
-
-#Import Libraries
-import uproot
-import numpy as np
-import matplotlib.pyplot as plt
-import matplotlib.patches as mpatch
-plt.style.use('standard.mplstyle')
-from matplotlib.lines import Line2D
-#Custom libraries
-import sys
-sys.path.append('./python')
-import nc_kinematics as nck
-import lindhard as lin
-import R68_yield as R68y
-from hist import histogramable as h
-
-#Build stuff!
-
-#Select a file.
-file = './data/file.root'
-
-real_Lind = np.ndarray.flatten(np.asarray(h(file)[0]))
-real_Sor = np.ndarray.flatten(np.asarray(h(file,model='Sorenson')[0]))
-small_Lind = np.ndarray.flatten(np.asarray(h(file,scalefactor=0.2)[0]))
-small_Sor = np.ndarray.flatten(np.asarray(h(file,model='Sorenson',scalefactor=0.2)[0]))
-
-real_Lind = real_Lind[real_Lind >= 0]
-real_Sor = real_Sor[real_Sor >= 0]
-small_Lind = small_Lind[small_Lind >= 0]
-small_Sor = small_Sor[small_Sor >= 0]
-
-#From https://stackoverflow.com/questions/31517156/adjust-exponent-text-after-setting-scientific-limits-on-matplotlib-axis
-
-def format_exponent(ax, axis='y'):
-
-    # Change the ticklabel format to scientific format
-    ax.ticklabel_format(axis=axis, style='sci', scilimits=(-2, 2))
-
-    # Get the appropriate axis
-    if axis == 'y':
-        ax_axis = ax.yaxis
-        x_pos = 0.0
-        y_pos = 1.0
-        horizontalalignment='left'
-        verticalalignment='bottom'
-    else:
-        ax_axis = ax.xaxis
-        x_pos = 1.0
-        y_pos = -0.05
-        horizontalalignment='right'
-        verticalalignment='top'
-
-    # Run plt.tight_layout() because otherwise the offset text doesn't update
-    plt.tight_layout()
-
[…]
