### 2021-05-18 2cbebe7
Adjust documentation for removal of extra python script
