+++ example-usecase/requirements.txt
@@ -5,4 +5,5 @@ ipykernel
 pytest
 uproot==3.12.0
 papermill
+#Only really needed for testing - user can comment out but doesn't hurt anything
 nbconvert
