### 2021-05-20 da45a0e
Add md5 checksum for nrcascadesim output
