### 2021-05-20 ffbcb1a
Add checksum print to travis for troubleshooting.
