### 2021-05-20 efb7353
Update .travis.yml to checksum against current seeding method. Need to look into how much this actually changed - not sure why the checksum is any different, will take a closer look later
