+++ bin/realizeCascades.cpp
@@ -184,6 +184,7 @@ int main(int argc, char** argv) {
   cout << "Seed used: " << cl << endl;
 
   MTRand *mtrand = new MTRand(cl);
+  cout << "MTRand: " << mtrand << endl;
 
   //get a root file and make 
   TFile *f = new TFile(outputfile.c_str(),"recreate");
