### 2021-05-21 3f90b2c
Move most of stdout to optional text file logging.
