### 2021-05-21 9a058a2
Fix(?) syntax on check for emtpy string
