### 2021-05-21 6789428
Reinclude md5sum output to troubleshoot.
