+++ .travis.yml
@@ -45,5 +45,6 @@ script:
 - travis_wait 50 py.test
 - cd ../
 - bin/realizeCascades -n 100000 -o "output.root?reproducible" -d 1 levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
+- md5sum output.root
 - md5sum output.root | awk '{if ($1=="cafd595aa3e650b317fd8c72ed60d963"){print"Checksum passed!"} else {exit 1}}'
 # NOTE that the checksum seems for some reason to be dependent on the relative path given
