### 2021-05-21 07da5cd
Change md5sum check value for CI Also remove --debug from conda install as it is no longer needed
