+++ .travis.yml
@@ -14,7 +14,7 @@ install:
   - conda info -a
 
   #- conda config --set channel_priority strict
-  - conda env create --debug -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
+  - conda env create -n test-environment python=$TRAVIS_PYTHON_VERSION -f nrCascadeSim_env.yml 
   - source activate test-environment
 
   # Snap installation of root
@@ -45,6 +45,6 @@ script:
 - travis_wait 50 py.test
 - cd ../
 - bin/realizeCascades -n 100000 -o "output.root?reproducible" -d 1 levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
-- md5sum output.root
-- md5sum output.root | awk '{if ($1=="cafd595aa3e650b317fd8c72ed60d963"){print"Checksum passed!"} else {exit 1}}'
+#- md5sum output.root
+- md5sum output.root | awk '{if ($1=="31925f4d132276ae792930646b76c2c5"){print"Checksum passed!"} else {exit 1}}'
 # NOTE that the checksum seems for some reason to be dependent on the relative path given
