+++ README.md
@@ -59,9 +59,21 @@ Making the full example: `./realizeCascades -n 100000 -o ~/output.root levelfile
 
 (More detailed examples below.)
 
+### Reproducible Files
+If you want a reproducible file that can be compared to another by an md5 checksum, append to the output file's name:
+`?reproducible=fixedname` (you will either need to put the filename in quotes or escape the `?` character).
+
+Example: `-o "output.root?reproducible=fixedname"`
+
+This surpresses various forms of metadata that result in changes to the binary even for the same data.
+(See [ROOT's page on the TFile class](https://root.cern.ch/doc/master/classTFile.html#ad0377adf2f3d88da1a1f77256a140d60).)
+
+If you wish to verify the program, the md5sum of a file generated by the following command:
+`bin/realizeCascades -n 100000 -o "output.root?reproducible=fixedname" -d 1 levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt`
+should be:
+`4fdf2ca6a36f218e906294082d51d800`
+
 ## Examples
-
-
 
 There is also a directory `example-usecase` containing one example of how data can be used once generated in a jupyter notebook `Yields_and_Resolutions.ipynb`. 
 This notebook serves to help visualize what the final data can look like as well as provide a few examples of how the data in the output root file can be accessed.
