### 2021-05-21 b210a1e
Switch md5sum check again. Doesn't seem possible to get a one-size-fits-all; must be dependent on something in the environment? (Possibly ROOT minor version?)
