+++ README.md
@@ -60,18 +60,18 @@ Making the full example: `./realizeCascades -n 100000 -o ~/output.root levelfile
 (More detailed examples below.)
 
 ### Reproducible Files
-If you want a reproducible file that can be compared to another by an md5 checksum, append to the output file's name:
-`?reproducible=fixedname` (you will either need to put the filename in quotes or escape the `?` character).
+The `-d` flag for the seed will result in files with consistent data. 
 
-Example: `-o "output.root?reproducible=fixedname"`
-
+However, additional binary data may result in checksums being different despite the data being the same. 
+If you want a reproducible file that can be compared to another by an md5 checksum, append to the output file's name:  
+`?reproducible=fixedname`  
+(you will either need to put the filename in quotes or escape the `?` character).  
+Example: `-o "output.root?reproducible=fixedname"`  
 This surpresses various forms of metadata that result in changes to the binary even for the same data.
 (See [ROOT's page on the TFile class](https://root.cern.ch/doc/master/classTFile.html#ad0377adf2f3d88da1a1f77256a140d60).)
 
-If you wish to verify the program, the md5sum of a file generated by the following command:
-`bin/realizeCascades -n 100000 -o "output.root?reproducible=fixedname" -d 1 levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt`
-should be:
-`4fdf2ca6a36f218e906294082d51d800`
+Unfortunately, the binary still seems to be influenced by the environment it is generated in,
+so at present there is no md5sum to compare to.
 
 ## Examples
 
