+++ .travis.yml
@@ -46,5 +46,5 @@ script:
 - cd ../
 - bin/realizeCascades -n 100000 -o "output.root?reproducible=fixedname" -d 1 levelfiles/Si28_ngam_all_cascades_rfmt_sorted.txt
 - md5sum output.root
-- md5sum output.root | awk '{if ($1=="4fdf2ca6a36f218e906294082d51d800"){print"Checksum passed!"} else {exit 1}}'
+- md5sum output.root | awk '{if ($1=="2029e8b402b89d3a05d745e48f546370"){print"Checksum passed!"} else {exit 1}}'
 # NOTE that the checksum seems for some reason to be dependent on the relative path given
