### 2021-05-21 6a24cf7
Revert "Switch ROOT to 6.22.00 (from 6.22.08) to match local machine."
This reverts commit caa5dbf800469fc133797dee58ce990a2c1805ba.
Changing ROOT version did not result in a match in checksums,
So changing things back to get back to a passing version.
