+++ .travis.yml
@@ -32,8 +32,8 @@ install:
 #  - conda install root
 
   # Get ROOT via binary
-  - wget https://root.cern/download/root_v6.22.00.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
-  - tar -xzvf root_v6.22.00.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
+  - wget https://root.cern/download/root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
+  - tar -xzvf root_v6.22.08.Linux-ubuntu16-x86_64-gcc5.4.tar.gz
   - sudo apt-get install make cmake g++ gcc binutils libx11-dev libxpm-dev libxft-dev libxext-dev openssl libssl-dev
   - source root/bin/thisroot.sh
 
