### 2021-05-22 3432817
Attempt to address comments on lines 10-12
