### 2021-05-22 82a9c92
Move figure 1 from paper to README to address part of comments on summary.
