### 2021-05-22 620cfad
Reduce verbosity of figure caption.
