+++ paper/paper.md
@@ -98,13 +98,13 @@ neutrino detection (CE$\mathrm{\nu}$NS).
 # Example Use Case
 
 Included in the repository is an example `test-example/Yields_and_Resolutions.ipynb` which users
-can follow to ensure the code is running correctly.  This example both applies a yield model to
-the individual energy deposits and applies variation intended to simulate the resolution of the
-detector.  The yield and resolution models are described in more detail in the example notebook.
+can follow to ensure the code is running correctly.  This example both applies multiple yield models &mdash; Lindhard [@lindhard] and Sorensen [@sorensen] &mdash; to
+the individual energy deposits and applies variations intended to simulate the resolution of the
+detector.  It also normalizes the frequencies for each bin of energy yields. The yield and resolution models are described in more detail in the example notebook.
 Figure \ref{LindvSor_fig} shows overlaid histograms of different combinations of analysis on the same
 data file.
 
-![An overlaid histogram showing an example use case in which points are generated and then multiple yield models and resolutions are applied.  In this example, the x-axis represents the ionization energy "yielded" by the cascade; this is effectively a way of noting what the detector reads out as opposed to what the pure kinetic energy of the cascade is.  The Lindhard yield [@lindhard] is output by \texttt{nrCascadeSim} as \texttt{Ei}; the Sorenson yield [@sorensen] is applied to the values from \texttt{delE}. Resolutions are applied by adding random values generated from a Gaussian distribution of fixed width to the energy yield.  The "Small Res (1/5)" histograms have Gaussians with 1/5 of the width of their counterparts.  The y-axis represents the normalized frequency of energy yields.\label{LindvSor_fig}](SorVsLin_fig.pdf)
+![An overlaid histogram showing an example use case in which points are generated and then multiple yield models and resolutions are applied.  The "Small Res (1/5)" histograms have Gaussians with 1/5 of the width of their counterparts. \label{LindvSor_fig}](SorVsLin_fig.pdf)
 
 # Acknowledgements
 
