### 2021-05-28 6e7edf5
Move Lindhard equations from paper to documentation.
