### 2021-05-28 8223fd6
Fix format of output_structure image
