### 2021-05-28 85931fa
Remove example usecase section, reduce explanation of figure, move figure.
