### 2021-05-29 e18a429
Add state of field (included in statement of need) Discuss GEANT4 and what we do that it doesn't No other packages known that do this kind of simulation.
