### 2021-05-29 dfb19cc
Add CI to standards.
