### 2021-05-29 14571ed
Add link to Travis-CI docs for convenience.
