### 2021-06-02 ee1da46
Improve formatting of Levelfile information and fix confusing or conflicting statements.
