+++ README.md
@@ -68,68 +68,112 @@ The processed data in this notebook is an example of what might be used for neut
 # Levelfile (Input) Format
 
 The levelfile is a singular text file read by the program using regular expressions.
-Each row in a levelfile corresponds to one cascade, 
+Each row in a levelfile corresponds to one possible cascade, 
 which should include a relative weight for the probability of the cascade's occurence.
 While it is helfpul to create columns that are easy for the user to read,
 columns can be delineated by any number of spaces.
 
 The general format of one row of an input file is:
 
-weight Sn [..,E2,E1,E0] [..,tau2,tau1,inf]
+weight isotope A [..,E2,E1,E0] [..,tau2,tau1,inf]
 
 Each portion of this row is described below.
 
-## Brief Descriptions
+| `Name` | `Format` | Description |
+| ---   | ---   | --- |
+|`weight` | `#.##` or `#e+/-##` | The probability of this cascade occuring, normalized to unity with all other cascades. This variable includes a weight for the isotope's abundance compared to other isotopes listed within the same levelfile. (If only one isotope is present within the levelfile, the abundance weight is not needed.) Weights can be given in decimal form or scientific notation (e.g. 0.000671 or 6.71e-04). |
+| `isotope` | `##Xx` | The isotope of the nucleus *after* capture. (For example, if 28Si is present, it will become 29Si, so 29Si should be listed.) This should be formatted as two numbers, one capital letter, and one lower-case letter (e.g. 29Si, 74Ge).|
+| `A` | `##` | The number of particles in the nucleus after capture. This should match the first two digits of `isotope`. For example, if `isotope` is 72Ge, `A` should be 72. |
+| `energies` | `[... E2 E1 0]`| An ordered list of the energy levels traversed (keV), including the ground state (0 keV), separated by spaces. These should be in the decreasing order, the order in which the nucleus will go through the states. Do not include the separation energy to account for the initial unbound state before capture; this is already assumed.|
+| `lifetimes` | `[... tau2 tau1 inf]` | An ordered list of the lifetimes of the energy levels traversed (as), separated by spaces. It must be the same length as the list of energies, and the lifetimes should be in the same order as the energies. The last entry is `100000000000000.0` (1e+14 as, or 1 ms), which is effectively infinite on the timescale of the simulation, to indicate that the state is stable at the ground state.|
 
-``weight``: weight of this cascade realization. It will be normalized to unity with all other cascades
+## On Weights
 
-``Sn``: neutron separation energy (MeV), can include several isotopes if energy different. 
+The sum of the probabilities must be less than or equal to one in order for the simulation to 
+work properly. If the sum is less than one, the simulation may skip generating some points in 
+the output &mdash; for example, when requesting 100 entries, if the total probability is 0.95, 
+one would expect 95 entries on average &mdash; but the input cascades will still be at the 
+correct proportions with respect to one another. If the sum is greater than one, the simulation 
+may not reach certain cascades at all &mdash; for instance, if a file has 12 cascades, and the 
+probabilities of the first 10 add up to 1, then the last two will never be generated.
 
-``[..,E2,E1,E0]``: ordered list of the energy levels traversed (keV), including the ground state
+### **An example for calculating weights:**
 
-``[..,tau2,tau1,inf]``: ordered list of the lifetimes of the energy levels traversed (fs), it must be
-the same length as the previous list.
+A detector has three isotopes, which become 29Si, 30Si, and 31Si after capture. 
[…]
