### 2021-06-02 c4a72c3
Remove placeholder section on unused feature
