### 2021-06-03 3f85abf
Update README to include date and version in accordance with new standards.
