### 2021-06-03 803c2eb
Fix issues with quiet and silent arguments that came from logging. Quiet and silent are no longer distinct, so quiet was removed.
