### 2021-06-04 #3174 @gerudo7
We are currently working on addressing @matthewfeickert 's comments and plan to release all of the changes together as v1.1.0. We had another question come up about one of the comments:

> I can see that there is a `example-usecase` directory in the repository that is never mentioned in the `README`, but there is no documentation file in the directory itself and the only form of documentation is the notebook, which is a nice visualization but not a demo.

I am having some trouble understanding what separates this visualization from a demo. I have some ideas as to what the issue might be (below), but I was hoping I could get some guidance to see if I'm looking in the right direction, so to speak.
* Sample data is provided, so the user doesn't have to use `realizeCascades` if they choose not to.
* Too much of the code is hidden in custom libraries, so the user doesn't see the intermediate steps involved in calculation.
* No intermediate calculations are visualized, so the user only really *sees* the end result.
* There aren't explicit instructions for changing the calculations, so the user is not encouraged to interact with the code.
