### 2021-06-05 a6e6331
Add reference for conference proceeding. Think this should be able to contain the paper reference as well.
