+++ paper/references.bib
@@ -21,7 +21,7 @@
     url = {https://www.nndc.bnl.gov/nudat2/}
 }
 
-@article{ROOT,
+@article{ROOT_paperonly,
     title = {ROOT - An Object Oriented Data Analysis Framework},
     author = {Rene Brun and Fons Rademakers},
     journal = {Nucl. Inst. & Meth. in Phys. Res. A},
@@ -30,6 +30,16 @@
     year = {1997},
     note = "See also: \url{https://root.cern/}",
     url = {https://root.cern/}
+}
+
+@inproceedings{ROOT,
+    title = {ROOT - An Object Oriented Data Analysis Framework},
+    author = {Rene Brun and Fons Rademakers},
+    maintitle = {Proceedings AIHENP'96 Workshop},
+    booktitle = {Nucl. Inst. & Meth. in Phys. Res. A 389 (1997) 81-86.},
+    year = {1996},
+    month = {Sep.},
+    note = {See also "ROOT" [software], Release v6.22/00, 02/07/2020, (No DOI available for this version.)}
 }
 
 @article{lindhard,
