### 2021-06-05 a88eb94
Force note to show up by marking it as a url.
