+++ paper/references.bib
@@ -39,7 +39,7 @@
     booktitle = {Nucl. Inst. & Meth. in Phys. Res. A 389 (1997) 81-86.},
     year = {1996},
     month = {Sep.},
-    note = {See also "ROOT" [software], Release v6.22/00, 02/07/2020, (No DOI available for this version.)}
+    url = {See also "ROOT" [software], Release v6.22/00, 02/07/2020, (No DOI available for this version.)}
 }
 
 @article{lindhard,
