+++ paper/references.bib
@@ -21,7 +21,7 @@
     url = {https://www.nndc.bnl.gov/nudat2/}
 }
 
-@article{ROOT_paperonly,
+article{ROOT_paperonly,
     title = {ROOT - An Object Oriented Data Analysis Framework},
     author = {Rene Brun and Fons Rademakers},
     journal = {Nucl. Inst. & Meth. in Phys. Res. A},
@@ -39,8 +39,10 @@
     booktitle = {Nucl. Inst. & Meth. in Phys. Res. A 389 (1997) 81-86.},
     year = {1996},
     month = {Sep.},
-    url = {See also "ROOT" [software], Release v6.22/00, 02/07/2020, (No DOI available for this version.)}
+    note = {See also "ROOT" [software], Release v6.22/00, 02/07/2020, (No DOI available for this version.)},
+    url = {Nucl. Inst. & Meth. in Phys. Res. A 389 (1997) 81-86. See also "ROOT" [software], Release v6.22/00, 02/07/2020, (No DOI available for this version.)}
 }
+
 
 @article{lindhard,
     title = {Integral Equations Governing Radiation Effects},
