### 2021-06-05 villano-lab/nrCascadeSim#3 @gerudo7
I think this is good now as the README now provides possible applications of the program in addition to information about what the simulation is doing.
