### 2021-06-05 e103dea
Have `-v 2` or greater include the information that was moved to logfile.
