### 2021-06-09 #3174 @matthewfeickert
> I am having some trouble understanding what separates this visualization from a demo. I have some ideas as to what the issue might be (below), but I was hoping I could get some guidance to see if I'm looking in the right direction, so to speak.

I'm a bit short on time this week, so I can get more detailed later, but so that you get some response sooner then later:

- Demonstrations are fully runnable examples that are showing a user how one interacts with the grammar of the API and a potential use workflow. They ideally exist in as a mixture of both code as well as through annotation and discussion on the purpose (here the physics) or what is happening.
- Visualizations are useful for showing what a possible data product of running code could be, but unless prohibitively computationally intensive/long they benefit greatly when used in documentation if they're fully reproducible.
