### 2021-06-30 #3174 @matthewfeickert
@mbobra @kelseymh @gerudo7 @villaa Just coming back to this before my July gets consumed by conferences. Can we have an update on the status of everything?

It seems from glancing at the project that planned revisions (which look good in terms of scope) have slowed and at the moment we only have my review (no blame or shame meant here &mdash; life happens and this year is a very exhausting one for everyone). The review has been going on for over 3 months now and I know that my time will be exceedingly restricted in July. Does it make sense to have this particular review continue for the time being? Or should we close this review and invite a resubmission once the authors are happy with their revisions?
