### 2021-06-30 #3174 @mbobra
I understand both of your concerns. On one hand, @matthewfeickert has provided an inordinate amount of constructive feedback, on his own volunteer time, for a period of months. This is more time than we at JOSS want to ask of our reviewers. On the other hand, @villaa have made progress and worked to improve the code.

So I am going to pause the review. This gives the submitting authors time to improve their code and also gives reviewers @matthewfeickert and @kelseymh a break from keeping up with this review thread.

@villaa Please message me when you feel your software is ready. At that time, I'll ask @matthewfeickert and @kelseymh to take a final look and give their final yay or nay. Then we will either accept or reject this submission. If @matthewfeickert and/or @kelseymh are not available at that time, I will assign a new reviewer (or two).

Thanks all for your hard work -- I really appreciate it.
