### 2021-07-01 #3174 @matthewfeickert
Thanks @mbobra &mdash; this all sounds reasonable to me. :+1:

> another possibility is that we can commit to having all of your comments fully addressed by the end of July and then you can finish the review out when you're back. Does that work?

@villaa @gerudo7 Sure. I'm not going to ask you to commit to a timeline here as I know the summer is a very busy time for everyone. So whenever things are at a state that you're happy with on your side please just ping @mbobra and me here and we'll be quite happy to resume then. :+1: I'm now marking this Issue as "Done" in my GitHub notifications so that it gets cleared off, but it will come back up the next time there's any activity on this Issue.

> I kind of feel that sometimes reviews go for a long-ish time and asking us to resubmit now is unfair.

I don't know if it is any consolation, but my suggestion was made out of a concern of fairness towards you and your time. I'm projecting my own feelings from past experiences in academic publishing here, but long running reviews that are met with further requests for revisions are their own special sort of terrible and don't motivates feelings of a shared goal between authors and reviewers. I'd like your experience with JOSS to be as positive as possible!
