### 2021-07-13 0087df6
Fix Eg units
They didn't seem to be eV based on output. Comment in cascadeProd.h indicates they are MeV.
