### 2021-09-30 #3174 @matthewfeickert
> So, I _think_ we can address all the issues raised in the week after that.

Can you clarify this more? There has been no activity on the repository for well over a month

![activity](https://user-images.githubusercontent.com/5142394/135511775-f11474c7-b6e6-4fd1-a229-4e53e626433f.png)

I will be honest and say that I'm don't see how this review will be done in a reasonable time. The plan was that the authors would take the summer to revise and prepare the software for review and this did not happen. I say this with no judgement as I think it is safe to assume that everyone has had a very busy summer for a multitude of reasons &mdash; I know I have.

> I don't know if our reviewers are willing to re-engage at that time.

@danielskatz I think that this review should be closed and if the authors want to submit again they are welcome to. I am the only reviewer who has participated and so another reviewer would need to be found for this to even get to an approval stage. It is not fair to anyone involved in the process to have a review that started in April be continuing at this point if it is not in the final stages.
