### 2021-09-30 #3174 @mbobra
@villaa I am the editor and responsible for removing and assigning reviewers. Resubmitting doesn't actually waste any time. If the repo is almost finished, then the new reviewers will be able to go through it quite easily and give it an overall yay or nay.

@matthewfeickert has indeed spent quite a bit of time watching this. I can't in good faith ask him to continue on his volunteer time to monitor a review for half a year. So I personally think you might want to consider re-submitting. It is a win-win all around -- you get as much time as you need to polish the repository, and when you're done the review process won't take multiple months.
